+++ pyproject.toml
@@ -79,6 +79,7 @@ sql_postgres = [
 ]
 stagehand = [
     "stagehand>=3.19.0",
+    "playwright>=1.27.0",
 ]
 sql_sqlite = [
     "sqlalchemy[asyncio]>=2.0.0,<3.0.0",
